revise: resnet to parallel
